Rename cloudflare package into terraform.
