+++ cf-ops-automation-terraform-broker/src/test/java/com/orange/oss/cloudfoundry/broker/opsautomation/ondemandbroker/terraform/TerraformServiceProvisionningTest.java
@@ -1,8 +1,7 @@
-package com.orange.oss.cloudfoundry.broker.opsautomation.ondemandbroker.cloudflare;
+package com.orange.oss.cloudfoundry.broker.opsautomation.ondemandbroker.terraform;
 
 
 import com.orange.oss.cloudfoundry.broker.opsautomation.ondemandbroker.git.GitServer;
-import com.orange.oss.cloudfoundry.broker.opsautomation.ondemandbroker.terraform.TerraformModuleHelper;
 import com.orange.oss.cloudfoundry.broker.opsautomation.ondemandbroker.terraform.cloudflare.TerraformProperties;
 import io.restassured.RestAssured;
 import org.apache.http.HttpStatus;
